Rename package to timer
