+++ src/OptionsApp.vue
@@ -7,8 +7,8 @@ import { onMounted, ref } from 'vue'
 import { ChromeCommunicationManager } from './infra/chrome/communication'
 import { DailyResetTimeStorageService } from './domain/daily_reset_time/storage'
 import { ReloadService } from './infra/chrome/reload'
-import { FocusSessionRecordStorageService } from './domain/pomodoro/record/storage'
-import { TimerConfigStorageService } from './domain/pomodoro/config/storage'
+import { FocusSessionRecordStorageService } from './domain/timer/record/storage'
+import { TimerConfigStorageService } from './domain/timer/config/storage'
 import { CurrentDateService } from './infra/current_date'
 import { NotificationSettingStorageService } from './domain/notification_setting/storage'
