+++ src/domain/timer/config/index.spec.ts
