+++ src/domain/timer/config/index.ts
