+++ src/domain/timer/config/serialize.ts
