+++ src/domain/timer/duration.spec.ts
