+++ src/domain/timer/duration.ts
