+++ src/domain/timer/stage.ts
