+++ src/domain/timer/state/index.ts
