+++ src/domain/timer/state/serialize.ts
