+++ src/domain/timer/timer.spec.ts
