+++ src/pages/BlockingSettingPage/TimerIntegrationSetting.spec.ts
@@ -3,8 +3,8 @@ import { describe, expect, it } from 'vitest'
 import config from '../../config'
 import type { BlockingTimerIntegration } from '../../domain/blocking_timer_integration'
 import { BrowsingRules } from '../../domain/browsing_rules'
-import { TimerStage } from '../../domain/pomodoro/stage'
-import { newTestTimerState } from '../../domain/pomodoro/state'
+import { TimerStage } from '../../domain/timer/stage'
+import { newTestTimerState } from '../../domain/timer/state'
 import { FakeActionService } from '../../infra/action'
 import { assertSelectorCheckboxValue } from '../../test_utils/assert'
 import { setUpListener } from '../../test_utils/listener'
