+++ src/pages/ReminderPage.spec.ts
@@ -1,10 +1,10 @@
 import { flushPromises, mount } from '@vue/test-utils'
 import { describe, expect, it } from 'vitest'
 import { DailyResetTimeStorageService } from '../domain/daily_reset_time/storage'
-import { TimerConfig } from '../domain/pomodoro/config'
-import { Duration } from '../domain/pomodoro/duration'
-import { newFocusSessionRecord } from '../domain/pomodoro/record'
-import { TimerStage } from '../domain/pomodoro/stage'
+import { TimerConfig } from '../domain/timer/config'
+import { Duration } from '../domain/timer/duration'
+import { newFocusSessionRecord } from '../domain/timer/record'
+import { TimerStage } from '../domain/timer/stage'
 import { Time } from '../domain/time'
 import { FakeActionService } from '../infra/action'
 import { CurrentDateService } from '../infra/current_date'
