+++ src/pages/ReminderPage.vue
@@ -3,8 +3,8 @@ import type { CurrentDateService } from '@/infra/current_date'
 import type { ClientPort } from '@/service_workers/listener'
 import { computed, onBeforeMount, ref } from 'vue'
 import type { DailyResetTimeStorageService } from '../domain/daily_reset_time/storage'
-import type { FocusSessionRecordStorageService } from '../domain/pomodoro/record/storage'
-import { TimerStage } from '../domain/pomodoro/stage'
+import type { FocusSessionRecordStorageService } from '../domain/timer/record/storage'
+import { TimerStage } from '../domain/timer/stage'
 import { Time } from '../domain/time'
 import { WorkRequestName } from '../service_workers/request'
 import { WorkResponseName } from '../service_workers/response'
