+++ src/pages/StatisticsPage.spec.ts
@@ -1,9 +1,9 @@
 import { flushPromises, mount, VueWrapper } from '@vue/test-utils'
 import { describe, expect, it } from 'vitest'
 import { DailyResetTimeStorageService } from '../domain/daily_reset_time/storage'
-import { TimerConfig } from '../domain/pomodoro/config'
-import { Duration } from '../domain/pomodoro/duration'
-import { newFocusSessionRecord, type FocusSessionRecord } from '../domain/pomodoro/record'
+import { TimerConfig } from '../domain/timer/config'
+import { Duration } from '../domain/timer/duration'
+import { newFocusSessionRecord, type FocusSessionRecord } from '../domain/timer/record'
 import { Time } from '../domain/time'
 import { FakeActionService } from '../infra/action'
 import { CurrentDateService } from '../infra/current_date'
