+++ src/pages/StatisticsPage.vue
@@ -3,7 +3,7 @@ import type { ReloadService } from '@/infra/chrome/reload'
 import type { ClientPort } from '@/service_workers/listener'
 import { onBeforeMount, ref } from 'vue'
 import { DailyResetTimeStorageService } from '../domain/daily_reset_time/storage'
-import type { FocusSessionRecordStorageService } from '../domain/pomodoro/record/storage'
+import type { FocusSessionRecordStorageService } from '../domain/timer/record/storage'
 import { Time } from '../domain/time'
 import { CurrentDateService } from '../infra/current_date'
 import { WorkRequestName } from '../service_workers/request'
