+++ src/pages/TestingConfigPage.spec.ts
@@ -1,8 +1,8 @@
 import { describe, expect, it } from 'vitest'
-import { TimerConfig } from '../domain/pomodoro/config'
+import { TimerConfig } from '../domain/timer/config'
 import { flushPromises, mount } from '@vue/test-utils'
 import TestingConfigPage from './TestingConfigPage.vue'
-import { Duration } from '../domain/pomodoro/duration'
+import { Duration } from '../domain/timer/duration'
 import { setUpListener } from '../test_utils/listener'
 import { assertSelectorInputValue } from '../test_utils/assert'
 import { dataTestSelector } from '../test_utils/selector'
