+++ src/pages/TestingConfigPage.vue
@@ -1,9 +1,9 @@
 <script setup lang="ts">
 import type { ClientPort } from '@/service_workers/listener'
 import { defineProps, onBeforeMount, ref } from 'vue'
-import { TimerConfig } from '../domain/pomodoro/config'
-import { TimerConfigStorageService } from '../domain/pomodoro/config/storage'
-import { Duration } from '../domain/pomodoro/duration'
+import { TimerConfig } from '../domain/timer/config'
+import { TimerConfigStorageService } from '../domain/timer/config/storage'
+import { Duration } from '../domain/timer/duration'
 import { WorkRequestName } from '../service_workers/request'
 
 const { timerConfigStorageService, port } = defineProps<{
