+++ src/pages/TimerSettingPage.spec.ts
@@ -1,7 +1,7 @@
 import { flushPromises, mount } from '@vue/test-utils'
 import { describe, expect, it } from 'vitest'
-import { TimerConfig } from '../domain/pomodoro/config'
-import { Duration } from '../domain/pomodoro/duration'
+import { TimerConfig } from '../domain/timer/config'
+import { Duration } from '../domain/timer/duration'
 import { FakeActionService } from '../infra/action'
 import { assertSelectorCheckboxValue, assertSelectorInputValue } from '../test_utils/assert'
 import { setUpListener } from '../test_utils/listener'
