+++ src/service_workers/listener.ts
@@ -10,14 +10,14 @@ import type { BrowsingControlService } from '../infra/browsing_control'
 import { BrowsingControlTogglingService } from '../domain/browsing_control_toggling'
 import { BrowsingRulesStorageService } from '../domain/browsing_rules/storage'
 import { NotificationSettingStorageService } from '../domain/notification_setting/storage'
-import { TimerConfigStorageService } from '../domain/pomodoro/config/storage'
-import { newFocusSessionRecord } from '../domain/pomodoro/record'
-import { FocusSessionRecordHousekeeper } from '../domain/pomodoro/record/house_keep'
-import { FocusSessionRecordStorageService } from '../domain/pomodoro/record/storage'
-import { TimerStage } from '../domain/pomodoro/stage'
-import type { TimerState } from '../domain/pomodoro/state'
-import { TimerStateStorageService } from '../domain/pomodoro/state/storage'
-import { FocusTimer } from '../domain/pomodoro/timer'
+import { TimerConfigStorageService } from '../domain/timer/config/storage'
+import { newFocusSessionRecord } from '../domain/timer/record'
+import { FocusSessionRecordHousekeeper } from '../domain/timer/record/house_keep'
+import { FocusSessionRecordStorageService } from '../domain/timer/record/storage'
+import { TimerStage } from '../domain/timer/stage'
+import type { TimerState } from '../domain/timer/state'
+import { TimerStateStorageService } from '../domain/timer/state/storage'
+import { FocusTimer } from '../domain/timer/timer'
 import { WeeklyScheduleStorageService } from '../domain/schedules/storage'
 import { type ActionService } from '../infra/action'
 import { type BadgeColor, type BadgeDisplayService } from '../infra/badge'
